+++ client-library/package.json
@@ -4,6 +4,7 @@
   "main": "dist/index.js",
   "version": "0.1.0",
   "dependencies": {
+    "@biomejs/biome": "^1.7.3",
     "@neondatabase/serverless": "^0.9.3",
     "@swc/cli": "^0.3.12",
     "@swc/core": "^1.5.22",
@@ -18,6 +19,8 @@
   },
   "scripts": {
     "build": "cd src && swc . -d ../dist",
+    "lint": "biome lint ./src",
+    "format": "biome format ./src --write",
     "postinstall": "cd src && swc . -d ../dist"
   }
 }
